perf: remove [@inline] on varint, but have it inlined elsewhere
[@inline] might bloat encoder/decoders that have a lot of varints, so we
should avoid that. However, in the many cases where we read/write an
`int` and not an `int64` (including the length-prefix for bytes), we can
inline `varint` in the surrouding function so that the `int64` will not
be boxed just before/after conversion with `int`.

This way, reading/writing an int might still be a function call, but
will not require boxing of a int64 where an int suffices.