wip: in benchmarks, compare current `nested` vs a backward-writing system
with a custom buffer where we write from the end, it's possible to
write nested records without intermediate buffers and copies. One just
writes the sub-record, then its size (which means the size precedes it
in the actual byte slice). this is somewhat tricky though, and doesn't
seem meaningfully faster.